+++ skeleton/settings_qa_basic_site.py
@@ -1,8 +1,4 @@
-from foundry import settings as foundry_settings
-
 from skeleton.settings_qa_base import *
 
 
 SITE_ID = 2
-
-foundry_settings.compute_settings(sys.modules[__name__])